refactor(tk): move custom tk modules to view folder
