+++ noScribe/view/tk/__init__.py
@@ -1,0 +1,2 @@
+from . import CTkToolTips
+from . import tkHyperlinkManager
