+++ noScribe/view/tk/tkHyperlinkManager.py
